% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{darken_color}
\alias{darken_color}
\title{Darken colors}
\usage{
darken_color(hex, factor = 0.5)
}
\arguments{
\item{hex}{A hex color code.}

\item{factor}{A number between 0 and 1 for darkening.}
}
\value{
A darkened hex color code.
}
\description{
Darken colors
}
