% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dynamic_title}
\alias{get_dynamic_title}
\title{Dynamically adjust title properties based on text length}
\usage{
get_dynamic_title(title_text)
}
\arguments{
\item{title_text}{The raw string for the title.}
}
\value{
A list containing the wrapped text, calculated size, and vjust.
}
\description{
Dynamically adjust title properties based on text length
}
